% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_inundation.R
\name{calc_inundation}
\alias{calc_inundation}
\title{Calculate number of inundation days}
\usage{
calc_inundation()
}
\value{
data.frame of dayflow data
}
\description{
Downloads all available dayflow data and Fremont weir data, and calculates
the number of inundation days and whether inundation is present. Returns a
data frame of daily data from 1984-present.
}
\examples{
\dontrun{

inun <- calc_inundation()

}
}
