% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_summary.R
\name{calc_summary}
\alias{calc_summary}
\title{Calculate summary of inundation data}
\usage{
calc_summary()
}
\value{
data.frame A data frame of inundation summary data.
}
\description{
This gives summary information of the inundation data returned by \code{calc_inundation}.
Included are: max inundation days by water year, total inundation days per water year,
date of and flood peak value (max of dayflow) by water year,
first and last overtopping date by water year,
last day of flooding number of overtopping events
}
\examples{
summary <- calc_inundation()
}
