% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fre.R
\name{get_fre}
\alias{get_fre}
\title{Load Fremont weir Sacramento river height}
\usage{
get_fre(
  stationID = "FRE",
  start = "1940-01-01",
  end = as.character(Sys.Date())
)
}
\arguments{
\item{stationID}{Station identifier (see https://info.water.ca.gov/staMeta.html)}

\item{start}{Start date in YYYY-MM-DD}

\item{end}{End date in YYYY-MM-DD}
}
\value{
data.frame of Sacramento river heights
}
\description{
see here: "https://portal.edirepository.org/nis/dataviewer?packageid=edi.840.1&entityid=186964642c42e5a8b8e44fc87ff10bbf", (1998-2021 Yolo)
}
