% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_inundation.R
\name{calc_inundation}
\alias{calc_inundation}
\title{Calculate number of inundation days}
\usage{
calc_inundation()
}
\value{
data.frame of dayflow data
}
\description{
Calculate number of inundation days
}
