% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fre.R
\name{get_fre}
\alias{get_fre}
\title{Load Fremont weir Sacramento river height}
\usage{
get_fre(start = "1940-01-01", end = as.character(Sys.Date()))
}
\arguments{
\item{start}{Start date in YYYY-MM-DD}

\item{end}{End date in YYYY-MM-DD}
}
\value{
data.frame of Sacramento river heights
}
\description{
Download Fremont weir data giving Sacramento river height.
see documentation for \href{https://portal.edirepository.org/nis/mapbrowse?packageid=edi.840.1}{1998-2021 Yolo on EDI}
}
\examples{
\dontrun{

fre <- get_fre()

}
}
